consolidate sdpa processor
